+++ digiarch/commands/edit/common.py
@@ -51,6 +51,7 @@ def edit_file_value(
                 setattr(file, "lock", True)
             table.update(file)
             database.log.insert(event)
+            database.commit()
         event.log(INFO, *loggers, show_args=["uuid", "data"], path=file.relative_path)
 
 
@@ -70,8 +71,11 @@ def rollback_file_value(property_name: str) -> _RH:
             table = database.statutory_files
         else:
             return
+        if not isinstance(file, table.model):
+            raise TypeError(f"{type(file)} is not {table.model.__name__}")
         prev_value, next_value = event.data
         setattr(file, property_name, prev_value)
         table.update(file)
+        database.commit()
 
     return _handler