+++ digiarch/commands/identify.py
@@ -147,6 +147,7 @@ def identify_original_file(
     parent: UUID | None,
     original_path: str | PathLike | None,
     *loggers: Logger,
+    ignore_lock: bool = False,
 ):
     errors: list[Event] = []
     existing_file: OriginalFile | None = db.original_files[
@@ -154,8 +155,18 @@ def identify_original_file(
     ]
 
     if existing_file and not update:
+        Event.from_command(ctx, "skip", (existing_file.uuid, "original"), reason="exists").log(
+            INFO,
+            *loggers,
+            path=existing_file.relative_path,
+        )
         return
-    if existing_file and existing_file.lock:
+    if existing_file and existing_file.lock and not ignore_lock:
+        Event.from_command(ctx, "skip", (existing_file.uuid, "original"), reason="locked").log(
+            INFO,
+            *loggers,
+            path=existing_file.relative_path,
+        )
         return
 
     file = OriginalFile.from_file(siegfried_file.filename, avid.path, parent=parent)
@@ -177,7 +188,6 @@ def identify_original_file(
         )
 
     file.original_path = Path(original_path).relative_to(file.root) if original_path else file.relative_path
-    existing_file: OriginalFile | None = db.original_files[file]
 
     if existing_file:
         file.uuid = existing_file.uuid
@@ -200,7 +210,7 @@ def identify_original_file(
         db.log.insert(*errors)
 
     if update or not existing_file:
-        Event.from_command(ctx, "file", (file.uuid, "original")).log(
+        Event.from_command(ctx, "update" if existing_file else "new", (file.uuid, "original")).log(
             INFO,
             *loggers,
             puid=str(file.puid).ljust(10),
@@ -224,6 +234,7 @@ def identify_original_files(
     update: bool,
     parent: UUID | None,
     *loggers: Logger,
+    ignore_lock: bool = False,
 ):
     if not paths:
         return
@@ -240,6 +251,7 @@ def identify_original_files(
             parent,
             None,
             *loggers,
+            ignore_lock=ignore_lock,
         )
 
 
@@ -368,6 +380,7 @@ def grp_identify():
 )
 @option("--exclude", type=str, multiple=True, help="File and folder names to exclude.  [multiple]")
 @option("--batch-size", type=IntRange(1), default=100, show_default=True, help="Amount of files to identify at a time.")
+@option("--ignore-lock", is_flag=True, default=False, show_default=True, help="Re-identify locked files.")
 @option_dry_run()
 @pass_context
 def cmd_identify_original(
@@ -380,6 +393,7 @@ def cmd_identify_original(
     custom_signatures_file: str | None,
     exclude: tuple[str, ...],
     batch_size: int | None,
+    ignore_lock: bool,
     dry_run: bool,
 ):
     """
@@ -428,7 +442,10 @@ def cmd_identify_original(
                     bool(query),
                     None,
                     log_stdout,
+                    ignore_lock=ignore_lock,
                 )
+                if not dry_run:
+                    db.commit()
 
         end_program(ctx, db, exception, dry_run, log_file, log_stdout)
 
@@ -521,6 +538,8 @@ def cmd_identify_master(
             while batch := list(islice(files, batch_size)):
                 for sf_file in siegfried.identify(*batch).files:
                     identify_master_file(ctx, avid, db, sf_file, custom_signatures, actions, dry_run, log_stdout)
+                if not dry_run:
+                    db.commit()
 
         end_program(ctx, db, exception, dry_run, log_file, log_stdout)
 
@@ -596,6 +615,8 @@ def cmd_identify_access(
             while batch := list(islice(files, batch_size)):
                 for sf_file in siegfried.identify(*batch).files:
                     identify_converted_file(ctx, avid, db.access_files, "access", sf_file, dry_run, log_stdout)
+                if not dry_run:
+                    db.commit()
 
         end_program(ctx, db, exception, dry_run, log_file, log_stdout)
 
@@ -671,6 +692,8 @@ def cmd_identify_statutory(
             while batch := list(islice(files, batch_size)):
                 for sf_file in siegfried.identify(*batch).files:
                     identify_converted_file(ctx, avid, db.statutory_files, "statutory", sf_file, dry_run, log_stdout)
+                if not dry_run:
+                    db.commit()
 
         end_program(ctx, db, exception, dry_run, log_file, log_stdout)